Fixed non-working mysql database module
